+++ pkg/webui/console/components/webhook-form/index.js
@@ -27,7 +27,7 @@ import KeyValueMap from '../../../components/key-value-map'
 import ModalButton from '../../../components/button/modal-button'
 import WebhookFormatSelector from '../../containers/webhook-formats-select'
 import sharedMessages from '../../../lib/shared-messages'
-import { id as webhookIdRegexp } from '../../lib/regexp'
+import { id as webhookIdRegexp, apiKey as webhookAPIKeyRegexp } from '../../lib/regexp'
 import PropTypes from '../../../lib/prop-types'
 
 import { mapWebhookToFormValues, mapFormValuesToWebhook, blankValues } from './mapping'
@@ -46,6 +46,9 @@ const m = defineMessages({
   headersValuePlaceholder: 'Bearer my-auth-token',
   headersAdd: 'Add header entry',
   headersValidateRequired: 'All header entry values are required. Please remove empty entries.',
+  downlinkAPIKey: 'Downlink API Key',
+  downlinkAPIKeyDesc:
+    'The API Key will be provided to the endpoint using the X-Downlink-Apikey header',
 })
 
 const headerCheck = headers =>
@@ -64,6 +67,7 @@ const validationSchema = Yup.object().shape({
   base_url: Yup.string()
     .url(sharedMessages.validateUrl)
     .required(sharedMessages.validateRequired),
+  downlink_api_key: Yup.string().matches(webhookAPIKeyRegexp, sharedMessages.validateFormat),
 })
 
 @bind
@@ -150,6 +154,13 @@ export default class WebhookForm extends Component {
           placeholder="http://example.com/webhooks"
           component={Input}
           required
+        />
+        <Form.Field
+          name="downlink_api_key"
+          title={m.downlinkAPIKey}
+          component={Input}
+          description={m.downlinkAPIKeyDesc}
+          code
         />
         <Message component="h4" content={sharedMessages.messageTypes} />
         <Notification info content={m.messageInfo} small />